Skeleton of VersionRange class
Before much of development can be done, the basic set up is required to
work on a class. Unit tests and build system has been amended. The class
represents a range of acceptable versions. There are many ways to
define a range. In the most basic case its a set of acceptable versions.