build: Use pytest in the functional test suite
The unittest module is only taking us this far. `py.test` offers a simpler
format for testing as it allows functions to be used, out of the box
fixture level handling of command line arguments. Also, it is a more
structured output than the plain old unittest ever did.

* Replace the old unittest code with the fixtured test functions
* Replace the unittest assertions with the native assertion. py.test will
handle the rest.
* Invoke the pytest module directly in CMakeLists.txt. The problem here
is that the Python::Interpreter resolves to the physical location of the
interpreter, instead of the virtual environment. Which means that it cannot
pick up the modules installed in the virtual environment. Instead invoke
it directly. pytest is a decent test driver with all the output formatting
available, so there is no reason to use CTest.