refactor: Clean up the python scripts
Some bits of the python scripts for testing became unused due to the
change to pytest. The logging in particular is done through the
logging module from you can retrive it using the pytest configuration.

Other than that, a simple touch up using the PEP8 formatter for all
python scripts.