feat: Parse version as a variable
Instead of baking in the version number we want to make it something
that comes from the SCM instead and only gets baked into the build
artefacts.

Conan does not require the presence of a version number as it can be
supplied through the command line, though it is a bit tedious that way.